+++ bot/main/talium/inputs/Twitch4J/Twitch4JInput.java
@@ -11,9 +11,9 @@ import com.github.twitch4j.chat.events.channel.ChannelMessageEvent;
 import com.github.twitch4j.helix.TwitchHelix;
 import org.apache.commons.lang.RandomStringUtils;
 import talium.Registrar;
-import talium.inputs.shared.oauth.OAuthEndpoint;
-import talium.inputs.shared.oauth.OauthAccount;
-import talium.inputs.shared.oauth.OauthAccountRepo;
+import talium.oauthConnector.OAuthEndpoint;
+import talium.oauthConnector.OauthAccount;
+import talium.oauthConnector.OauthAccountRepo;
 import talium.system.Out;
 import talium.system.eventSystem.EventDispatcher;
 import talium.system.eventSystem.Subscriber;
